+++ shared/components/data/useData/v1.0.0/types.d.ts
@@ -5,6 +5,7 @@ export type Props = BaseReactProps & {
     fetchScheme: FetchScheme
     searchEnabled: boolean    
     searchString?: string
+    paginationEnabled: boolean
     paginationDbClass?: string
 }